+++ CHANGELOG.md
@@ -4,4 +4,5 @@ History of **rest2go** library with changes description.
 
 ## Next release
 
+- `FEATURE` Loaded once settings as YAML file
 - `FEATURE` GoLang library project initialization